refactor(reborn): isolate auth prompt enrichment
