+++ crates/ironclaw_reborn_composition/src/oauth_gate.rs
@@ -20,8 +20,8 @@ use ironclaw_turns::{TurnRunId, TurnScope};
 use secrecy::SecretString;
 use tokio::sync::Mutex as AsyncMutex;
 
+use crate::AuthChallengeView;
 use crate::input::OAuthClientConfig;
-use crate::projection::AuthChallengeView;
 
 const GATE_FLOW_TTL_SECONDS: i64 = 600;
