+++ crates/ironclaw_reborn_composition/src/projection/tests.rs
@@ -34,6 +34,8 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::time::Duration;
 use tokio::sync::Mutex;
 
+use crate::AuthChallengeView;
+
 mod cursor_validation;
 mod display_preview;
 mod failure_explanation;
