+++ crates/ironclaw_reborn_composition/src/slack_serve/e2e_auth_challenge.rs
@@ -1,0 +1,92 @@
+use std::sync::Mutex;
+
+use async_trait::async_trait;
+use ironclaw_auth::{AuthProductError, AuthProviderId, OAuthAuthorizationUrl};
+use ironclaw_host_api::{AgentId, ProjectId, UserId};
+use ironclaw_product_adapters::AuthPromptChallengeKind;
+use ironclaw_turns::{TurnRunId, TurnScope};
+
+use crate::{AuthChallengeProvider, AuthChallengeView};
+
+use super::{AGENT, AUTH_GATE, PROJECT, TENANT, USER};
+
+type AuthChallengeCall = (
+    TurnScope,
+    UserId,
+    TurnRunId,
+    String,
+    Vec<ironclaw_host_api::RuntimeCredentialAuthRequirement>,
+);
+
+#[derive(Debug, Default)]
+pub(super) struct FakeAuthChallengeProvider {
+    calls: Mutex<Vec<AuthChallengeCall>>,
+}
+
+impl FakeAuthChallengeProvider {
+    pub(super) fn assert_single_call(&self) {
+        let calls = self
+            .calls
+            .lock()
+            .unwrap_or_else(|poisoned| poisoned.into_inner())
+            .clone();
+        let [(scope, owner_user_id, run_id, gate_ref, credential_requirements)] = calls.as_slice()
+        else {
+            panic!(
+                "expected one auth challenge provider call, got {}",
+                calls.len()
+            );
+        };
+        assert_eq!(scope.tenant_id.as_str(), TENANT);
+        assert_eq!(scope.agent_id.as_ref().map(AgentId::as_str), Some(AGENT));
+        assert_eq!(
+            scope.project_id.as_ref().map(ProjectId::as_str),
+            Some(PROJECT)
+        );
+        assert_eq!(
+            scope.explicit_owner_user_id().map(UserId::as_str),
+            Some(USER)
+        );
+        assert_eq!(owner_user_id.as_str(), USER);
+        assert!(!run_id.to_string().is_empty());
+        assert_eq!(gate_ref, AUTH_GATE);
+        assert!(credential_requirements.is_empty());
+    }
+}
+
+#[async_trait]
+impl AuthChallengeProvider for FakeAuthChallengeProvider {
+    async fn challenge_for_gate(
+        &self,
+        scope: &TurnScope,
+        owner_user_id: &UserId,
+        run_id: TurnRunId,
+        gate_ref: &str,
+        credential_requirements: &[ironclaw_host_api::RuntimeCredentialAuthRequirement],
+    ) -> Result<Option<AuthChallengeView>, AuthProductError> {
+        self.calls
+            .lock()
+            .unwrap_or_else(|poisoned| poisoned.into_inner())
+            .push((
+                scope.clone(),
+                owner_user_id.clone(),
+                run_id,
+                gate_ref.to_string(),
+                credential_requirements.to_vec(),
+            ));
+        if owner_user_id.as_str() != USER || gate_ref != AUTH_GATE {
+            return Ok(None);
+        }
+        Ok(Some(AuthChallengeView {
+            kind: AuthPromptChallengeKind::OAuthUrl,
+            provider: AuthProviderId::new("provider".to_string())
+                .expect("static provider id should be valid"), // safety: static test provider id is valid.
+            account_label: None,
+            authorization_url: Some(
+                OAuthAuthorizationUrl::new("https://provider.example/oauth".to_string())
+                    .expect("static OAuth URL should be valid"), // safety: static test URL is valid.
+            ),
+            expires_at: None,
+        }))
+    }
+}
